# Performance improvements to set comparison for large sets.