--- a/Runtime/Assist/ValueRetrievers/NullableTimeSpanValueRetriever.cs
+++ b/Runtime/Assist/ValueRetrievers/NullableTimeSpanValueRetriever.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace TechTalk.SpecFlow.Assist.ValueRetrievers
 {