--- a/Runtime/Assist/ValueRetrievers/TimeSpanValueRetriever.cs
+++ b/Runtime/Assist/ValueRetrievers/TimeSpanValueRetriever.cs
@@ -17,7 +17,7 @@ namespace TechTalk.SpecFlow.Assist.ValueRetrievers
 
         public bool CanRetrieve(KeyValuePair<string, string> keyValuePair, Type targetType, Type propertyType)
         {
-            return propertyType == typeof(System.TimeSpan);
+            return propertyType == typeof(TimeSpan);
         }
     }
 }