Resubmitted as a performance improvement.

- As discussed, rocknet's performance improvements are much larger (4500 seconds to 23 seconds), but it involves a semantic change.

- These changes still improve the performance significantly, from 50 seconds to 27 seconds, but without any semantic change as all differences are still reported in the comparison exception. Attached is the test code (zipped linqpad query)
[MeasureSetComparison.zip](https://github.com/techtalk/SpecFlow/files/245128/MeasureSetComparison.zip)
. Note that I should serialize the test case to a file for repeatably and do multiple runs of course. I did 4 runs; two with each version.

Original:
Finished comparison in 54925 ms.
Finished comparison in 48328 ms.

This branch:
Finished comparison in 26380 ms.
Finished comparison in 28246 ms.

As you recommended I pushed the ToArray() call into FindAnyDifferences(). I like that style better too.

Thanks!