@darrencauthon 2016-05-02
I'm fine with this pull request.

I think all of the speed improvements are due to [this one area of code](https://github.com/techtalk/SpecFlow/pull/597/files#diff-101fd7ddc2f5a7c95ecff303c70cbaccR144), as the rest are code changes that will mostly get compiled to the same IL anyway.  

Tests pass, I say this is good-to-go.