@SabotageAndi 2016-05-03
Looks also good for me.
@gasparnagy can we merge it?