@darrencauthon 2016-05-03
@SabotageAndi Sorry... I don't mean to step on any toes or be presumptive about getting these merged.  I'm trying to not to let the pull requests from contributors sit around too long, as I have seen some frustration grow around this area.  Since these changes were all in the Assist namespace, which sort of is my area of specialty, and that area has very good test coverage, I slept on it.  Then merged. :smile: