@SabotageAndi 2016-05-03
No problem.
I am also for quick merges. It only gets more complicated with the time.
I was only surprised that GitHub did not refreshed me issue.